requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
attrs==20.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-ordering==0.6
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
